Fix support for big data packets.
- Video frames are big.

+++ wrapper/internal/implementations/wine/dllhost/callback_handler.go
@@ -19,7 +19,7 @@ func (c *CallbackHandler) HandleCallback(eventCode uint64, data []byte,
 	// TOOD(bga): Add error checking.
 	binary.Write(&b, binary.BigEndian, eventCode)
 	binary.Write(&b, binary.BigEndian, tag)
-	binary.Write(&b, binary.BigEndian, uint16(len(data)))
+	binary.Write(&b, binary.BigEndian, uint32(len(data)))
 	b.Write(data)
 
 	b.WriteTo(c.eventFile)
